# [Deepin-Kernel-SIG] [linux 6.6-y] [Upstream] Remove use of wait bookmarks