https://lore.kernel.org/all/20231010035829.544242-1-willy@infradead.org/T/#m25ba483bf902a43b2858751f9255e644b60e026a

## Summary by Sourcery

Remove support for wait bookmarks from the scheduler and filemap wakeup paths and simplify wakeup code accordingly.

Enhancements:
- Drop WAITQUEUE_WALK_BREAK_CNT and bookmark-related logic from wake queue walking.
- Simplify __wake_up_common and associated wrappers by removing bookmark parameters and loops.
- Remove __wake_up_locked_key_bookmark API and its uses in filemap.c.
- Eliminate WQ_FLAG_BOOKMARK definition and adjust wait queue flag values accordingly in wait.h