@opsiff 2025-07-03
Test result:
Before patch:
------------------------------------------------------------------------
Benchmark Run: 三 7月 02 2025 16:25:47 - 16:32:31
8 CPUs in system; running 1 parallel copy of tests

Dhrystone 2 using register variables       50655140.0 lps   (10.0 s, 1 samples)
Double-Precision Whetstone                     5024.0 MWIPS (10.0 s, 1 samples)
Execl Throughput                               4470.8 lps   (29.8 s, 1 samples)
File Copy 1024 bufsize 2000 maxblocks        706172.0 KBps  (30.0 s, 1 samples)
File Copy 256 bufsize 500 maxblocks          201359.0 KBps  (30.0 s, 1 samples)
File Copy 4096 bufsize 8000 maxblocks       2139543.0 KBps  (30.0 s, 1 samples)
Pipe Throughput                             1726969.6 lps   (10.0 s, 1 samples)
Pipe-based Context Switching                 136493.3 lps   (10.0 s, 1 samples)
Process Creation                               9053.5 lps   (30.0 s, 1 samples)
Shell Scripts (1 concurrent)                  12556.2 lpm   (60.0 s, 1 samples)
Shell Scripts (8 concurrent)                   4998.9 lpm   (60.0 s, 1 samples)
System Call Overhead                        1453198.2 lps   (10.0 s, 1 samples)

System Benchmarks Index Values               BASELINE       RESULT    INDEX
Dhrystone 2 using register variables         116700.0   50655140.0   4340.6
Double-Precision Whetstone                       55.0       5024.0    913.5
Execl Throughput                                 43.0       4470.8   1039.7
File Copy 1024 bufsize 2000 maxblocks          3960.0     706172.0   1783.3
File Copy 256 bufsize 500 maxblocks            1655.0     201359.0   1216.7
File Copy 4096 bufsize 8000 maxblocks          5800.0    2139543.0   3688.9
Pipe Throughput                               12440.0    1726969.6   1388.2
Pipe-based Context Switching                   4000.0     136493.3    341.2
Process Creation                                126.0       9053.5    718.5
Shell Scripts (1 concurrent)                     42.4      12556.2   2961.4
Shell Scripts (8 concurrent)                      6.0       4998.9   8331.6
System Call Overhead                          15000.0    1453198.2    968.8
                                                                   ========
System Benchmarks Index Score                                        1593.7

------------------------------------------------------------------------
Benchmark Run: 三 7月 02 2025 16:32:31 - 16:39:14
8 CPUs in system; running 8 parallel copies of tests

Dhrystone 2 using register variables      207894957.3 lps   (10.0 s, 1 samples)
Double-Precision Whetstone                    37085.2 MWIPS (9.4 s, 1 samples)
Execl Throughput                              24307.9 lps   (29.1 s, 1 samples)
File Copy 1024 bufsize 2000 maxblocks       3885591.0 KBps  (30.0 s, 1 samples)
File Copy 256 bufsize 500 maxblocks         1135295.0 KBps  (30.0 s, 1 samples)
File Copy 4096 bufsize 8000 maxblocks       8144126.0 KBps  (30.0 s, 1 samples)
Pipe Throughput                             9453691.4 lps   (10.0 s, 1 samples)
Pipe-based Context Switching                1137628.8 lps   (10.0 s, 1 samples)
Process Creation                              50709.8 lps   (30.0 s, 1 samples)
Shell Scripts (1 concurrent)                  43841.3 lpm   (60.0 s, 1 samples)
Shell Scripts (8 concurrent)                   5701.3 lpm   (60.0 s, 1 samples)
System Call Overhead                        9323152.0 lps   (10.0 s, 1 samples)

System Benchmarks Index Values               BASELINE       RESULT    INDEX
Dhrystone 2 using register variables         116700.0  207894957.3  17814.5
Double-Precision Whetstone                       55.0      37085.2   6742.8
Execl Throughput                                 43.0      24307.9   5653.0
File Copy 1024 bufsize 2000 maxblocks          3960.0    3885591.0   9812.1
File Copy 256 bufsize 500 maxblocks            1655.0    1135295.0   6859.8
File Copy 4096 bufsize 8000 maxblocks          5800.0    8144126.0  14041.6
Pipe Throughput                               12440.0    9453691.4   7599.4
Pipe-based Context Switching                   4000.0    1137628.8   2844.1
Process Creation                                126.0      50709.8   4024.6
Shell Scripts (1 concurrent)                     42.4      43841.3  10339.9
Shell Scripts (8 concurrent)                      6.0       5701.3   9502.1
System Call Overhead                          15000.0    9323152.0   6215.4
                                                                   ========
System Benchmarks Index Score                                        7546.0

After patch:
Dhrystone 2 using register variables       45735929.7 lps   (10.0 s, 1 samples)
Double-Precision Whetstone                     5026.5 MWIPS (10.0 s, 1 samples)
Execl Throughput                               4458.3 lps   (29.2 s, 1 samples)
File Copy 1024 bufsize 2000 maxblocks        711618.0 KBps  (30.0 s, 1 samples)
File Copy 256 bufsize 500 maxblocks          204417.0 KBps  (30.0 s, 1 samples)
File Copy 4096 bufsize 8000 maxblocks       2321920.0 KBps  (30.0 s, 1 samples)
Pipe Throughput                             1747502.7 lps   (10.0 s, 1 samples)
Pipe-based Context Switching                 136247.6 lps   (10.0 s, 1 samples)
Process Creation                               9092.3 lps   (30.0 s, 1 samples)
Shell Scripts (1 concurrent)                  12546.2 lpm   (60.0 s, 1 samples)
Shell Scripts (8 concurrent)                   5008.9 lpm   (60.0 s, 1 samples)
System Call Overhead                        1447375.3 lps   (10.0 s, 1 samples)

System Benchmarks Index Values               BASELINE       RESULT    INDEX
Dhrystone 2 using register variables         116700.0   45735929.7   3919.1
Double-Precision Whetstone                       55.0       5026.5    913.9
Execl Throughput                                 43.0       4458.3   1036.8
File Copy 1024 bufsize 2000 maxblocks          3960.0     711618.0   1797.0
File Copy 256 bufsize 500 maxblocks            1655.0     204417.0   1235.1
File Copy 4096 bufsize 8000 maxblocks          5800.0    2321920.0   4003.3
Pipe Throughput                               12440.0    1747502.7   1404.7
Pipe-based Context Switching                   4000.0     136247.6    340.6
Process Creation                                126.0       9092.3    721.6
Shell Scripts (1 concurrent)                     42.4      12546.2   2959.0
Shell Scripts (8 concurrent)                      6.0       5008.9   8348.2
System Call Overhead                          15000.0    1447375.3    964.9
                                                                   ========
System Benchmarks Index Score                                        1595.2

------------------------------------------------------------------------
Benchmark Run: 三 7月 02 2025 18:39:38 - 18:46:21
8 CPUs in system; running 8 parallel copies of tests

Dhrystone 2 using register variables      207875447.4 lps   (10.0 s, 1 samples)
Double-Precision Whetstone                    36511.1 MWIPS (10.0 s, 1 samples)
Execl Throughput                              24289.1 lps   (29.1 s, 1 samples)
File Copy 1024 bufsize 2000 maxblocks       4070364.0 KBps  (30.0 s, 1 samples)
File Copy 256 bufsize 500 maxblocks         1158825.0 KBps  (30.0 s, 1 samples)
File Copy 4096 bufsize 8000 maxblocks       8050024.0 KBps  (30.0 s, 1 samples)
Pipe Throughput                             9563225.7 lps   (10.0 s, 1 samples)
Pipe-based Context Switching                1167311.4 lps   (10.0 s, 1 samples)
Process Creation                              51302.4 lps   (30.0 s, 1 samples)
Shell Scripts (1 concurrent)                  43861.1 lpm   (60.0 s, 1 samples)
Shell Scripts (8 concurrent)                   5705.1 lpm   (60.0 s, 1 samples)
System Call Overhead                        9306447.0 lps   (10.0 s, 1 samples)

System Benchmarks Index Values               BASELINE       RESULT    INDEX
Dhrystone 2 using register variables         116700.0  207875447.4  17812.8
Double-Precision Whetstone                       55.0      36511.1   6638.4
Execl Throughput                                 43.0      24289.1   5648.6
File Copy 1024 bufsize 2000 maxblocks          3960.0    4070364.0  10278.7
File Copy 256 bufsize 500 maxblocks            1655.0    1158825.0   7002.0
File Copy 4096 bufsize 8000 maxblocks          5800.0    8050024.0  13879.4
Pipe Throughput                               12440.0    9563225.7   7687.5
Pipe-based Context Switching                   4000.0    1167311.4   2918.3
Process Creation                                126.0      51302.4   4071.6
Shell Scripts (1 concurrent)                     42.4      43861.1  10344.6
Shell Scripts (8 concurrent)                      6.0       5705.1   9508.5
System Call Overhead                          15000.0    9306447.0   6204.3
                                                                   ========
System Benchmarks Index Score                                        7601.0